+++ Changelog.md
@@ -4,6 +4,7 @@ Lilu Changelog
 - Allow loading on macOS 13 without `-lilubetaall`
 - Added Ventura dyld shared cache pathing
 - Changed SKL default ig-platform-id to KBL on macOS 13+
+- Added patch with masking support
 
 #### v1.6.0
 - Dropped internal shared patcher instance grabbing API